mime_alter: Don't disable --force-for-bad-html
The --force-for-bad-html option is supposed to allow you to have your
HTML text inserted even if there is no <body></body> elements in the
mail message.

However this option is overridden in AM_init() by resetting
glb.force_for_bad_html to 0 when building with -DALTERMIME_PRETEXT

Remove this initialisation from AM_init().

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>