mime_alter: Fix force inserting of post/pre HTML
By default alterMIME want's to insert the HTML either just after the
<body> or just before the </body> elements.

You can override this behaviour with the --force-for-bad-html option,
however even after the last fix, it was still broken.

Rather than re-writing lots of the code. Just add in a couple of checks
in AM_add_disclaimer_insert_html() to see if we have
--force-for-bad-html and that we didn't yet manage to insert any HTML
and then insert it. There's one for each of pre and post HTML.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>