mime_alter: Fix duplication of HTML disclaimer
There was a situation whereby the HTML disclaimer could end up being put
in twice.

e.g
    <br>
    <b>Goodbye</b>
    <br>
    </body>
    </html>
    <br>
    <b>Goodbye</b>
    <br>

This was due to the --force-for-bad-html option putting the HTML in
after the closing </html> tag.

This was down to a mixture of the recent work done to make this work and
a patch seemingly introduced back in 2004. In the end this patch seemed
to be doing more harm than good and simple removing said code makes only
one HTML disclaimer added in the right place.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>