b64: Fix Base64 handling
The insertion of Base64 encoded messages was failing for a number of
reasons.

For --pretext-b64= this was putting the filename of the encoded pretext
into the pretext_html_file variable not the pretext_b64_file variable.

For --disclaimer-b64= thia was not being processed due to a missing
check in altermime::main(). This was fixed in the previous commit
4fff606 ("altermime: Allow to use --disclaimer-{b64,html} on their own")

Then there was a bigger problem that would lead to a segmentation fault
in mime_alter::AM_insert_disclaimer_into_buffer(). This function takes a
decoded buffer of the message and should then prepend/append the decoded
insertion text to it. However this function was looking for the
insertion text in dd->disclaimer_text_plain when in fact it is in
dd->disclaimer_text_b64 when --pretext-b64= is used, however it _is_ in
dd->disclaimer_text_plain when --pretext= -force-into-b64 is used.

So we need to first look in dd->disclaimer_text_b64 then in
dd->disclaimer_text_plain for the insertion text.

That order will cater for things like --pretext= --pretext-b64=
--force-into-b64

Also in this function we would get the combined buffer truncated by two
characters. This was due to copying one less chararcter than what we had
allocated space for (total_space = message + insertion text + 1) and
also the inclusion of a newline (\n) in the combined output.

This was due to copying one less chararcter than what we had allocated
space for (total_space = message + insertion text + 1) because of using
total_size - 1 in the size argument to snprintf(3). I guess this was
thinking about the terminating null byte, however snprintf will write at
most size bytes including the null byte, so we can simply use
'total_size' (it already includes space for the null byte) as the size
argument.

The other character truncation was due to the including of a newline
(\n) character in the combined output so make total_size one bigger o
cater for this.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>