mime_alter: b64: Impove multipart MIME handling with base64 parts
In AM_insert_disclaimer_into_segment64() we need to check on
dd->b64_inserted instead of dd->text_inserted to see if we should add
the text otherwise if we've already added some plain text to a mime part
we won't be able to add anything to a base64 encoded part.

And we need to set dd->b64_inserted = 1 rather than dd->text_inserted.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>