mime_alter: Fix HTML insertion
If we inserted HTML text then AM_add_disclaimer_insert_html() would
proceed to read through to the end of file. This would mean that if we
wanted to insert into successive mime parts we couldn't as altermime
thought there was no more parts to insert into.

Fix this by doing the same thing that is done for inserting into
text/plain parts. As we read in each line check is we have reached the
mime part boundary, break out the loop and write the boundary to the new
file.

Also in AM_add_disclaimer() when reading in the last of the segments
don't use dd.html_inserted == 0 as a condition, we want to be able to
add HTML and then possibly others as well.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>