mime_alter: Don't try to force insert pre html
Looking at the code more, I don't think was intended to work and the
changes I made in commit e366d59 ("mime_alter: Fix force inserting of
post/pre HTML") would end up doing the wrong thing with html like

    <html>
    <body>
     ...
   </body>
   </html>

as it would see the <html> element and then do the force insertion above
it.

The only HTML that will be operated on correctly are the above and

    <body>
     ...
    </body>

So now if you use --force-for-bad-html it will only put in the
disclaimer part. Which seems to have been the orginally intended idea.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>