Revert "mime_alter: Don't try to force insert pre html"
This reverts commit ba5b83209fac154493e1dfb63048a3610025f448.

Client wants to force insert pre-html. Lets go with this simple solution
of if the first line in the mime part isnt <body> it will force insert
the pre-html.

Client says that most of the HTML email they need to process doesn't
have html or body taga, so this should generally be OK.