mime_alter: Prevent duplicate lines when force inserting pre-html
After commit 15b3a1c ("mime_alter: Fix HTML insertion") we could end up
with a repeated line in the output. e.g

    <br>
    <b>Hello World</b>
    <br>

    <pre>
    <pre>	<- duplicated
    Some html...
    </pre>

    <br>
    <b>Goodbye</b>
    <br>

This was due to adding in the last line read after force inserting the
pre-html.

This line was already going to be added by the final fprintf() which was
there to ensure that the mime part boundary got written.

So remove the fprintf() from the force insertion bit.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>