qpe: Fix line breaking of quoted printable text
When formatting quoted printable text, lines over 76 characters long
need to be soft breaked by putting a '=' at the end of the line.

There was code in qp_encode() to do this, but was ufortunately broken,
seemingly in an attempt to appease Outlook with comments like

    // Was 76, updated to fix Outlook problems

The main issue was that when putting a '=' at the end of the line to
signal a softbreak, it (and the line terminators) where subsequently
being overwritten and long lines where simply being encoded as long
lines.

The main culprit was this line

    op+= strlen(paragraph);// +3; /** jump the output + =\r\n **/

op needs to be set to the end of not just the paragrph but also the '='
and the line terminators.

The other issue was related to the top comment, the check had been
changed from >= 76 to >= 79, put it back the way it was. RFC 2045
clearly states 76 characters.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>