qpe: Fix erroneous insertion of CRLF's
When inserting disclaimers into messages it's important that any line
terminators being added are the same as is used throughout the message.

However, in qp_encode() CRLF was being added unconditionally to separate
paragraphs when it should be using the line_terminator variable.

Thus you could end up with a message that looked like (from the file
command)

    SMTP mail, ASCII text, with CRLF, LF line terminators

Notice that CRLF, that can lead to errors from mail servers along the
lines of

    550 Bare carriage returns not permitted

The mixture of CRLF and LF is probably causing the sending MTA to create
some duff line terminators when it does some line terminator replacement
to ensure that all line terminators are CRLF as mandated by RFC 2821.

After this fix we get

    SMTP mail, ASCII text

which doesn't cause any problems.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>