mime_alter: Silence a compiler warning
In AM_insert_HTML_disclaimer_into_segment64() we make a call to
truncate(2), this is is highly unlikely to fail. So simply get the
return status from this syscall and mark it __attribute__((unused)),
this prevents the following compiler warning

    cc -Wall -Werror -g -I. -O2 -Wp,-D_FORTIFY_SOURCE=2 -fstack-protector
    --param=ssp-buffer-size=4 -fPIC -DALTERMIME_PRETEXT -c mime_alter.c
    cc1: warnings being treated as errors
    mime_alter.c: In function ‘AM_insert_HTML_disclaimer_into_segment64’:
    mime_alter.c:2094: error: ignoring return value of ‘truncate’, declared
    with attribute warn_unused_result
    make: *** [mime_alter.o] Error 1

which occurs under the commit that follows due to new compiler flags.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>