mime_alter: Fix prehtml insertion
When using the --force-for-bad-html option the prehtml would often get
inserted right at the top just after the mime headers even if there was
a <body> tag further down.

This was due to AM_add_disclaimer_insert_html() just reading in the mime
part a line at a time and it seems more often than not that the <body>
won't just happen to be on the first line, where
AM_add_disclaimer_insert_html is looking for it and then proceeds to
insert it right at the start.

The simple fix is to just scan through the mime part first and see if
there is a <body> tag or not and only do the force insertion if the
--force-for-bad-html option is used *and* we don't have a <body> tag.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>