mime_alter: More prehtml insertion fixs
This continues the work on top of 593b724 ("mime_alter: Fix prehtml
insertion") and 025ba34 ("mime_alter: Fix truncated messages")

Inserting text immediately after the <body> tag seems to be an endless
source of issues. e.g the <body ...>  tag may itself be split across
multiple lines. Or it may be a few lines further down from where we
expect it.

This commit takes a slightly different approach in trying to find the
body tag. Basically while scanning through the HTML we are just looking
for '<body' and then a '>' if we find them we know we have a valid <body
...> tag. We can then go through the HTML again and insert the text at
the right location.

This simplifies the code a little and should make it more robust.

Signed-off-by: Andrew Clayton <andrew@zeta.digital-domain.net>